daemon: move datapath cell to pkg/datapath
Now that there are no more circular imports, we can move the datapath
cell and associated initialization code to pkg/datapath.

The sysctl stuff should most definitely not live toplevel here, but one
step at a time.
