+++ pkg/datapath/sysctl_darwin.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package datapath
 
 // enableIPForwarding on OS X and Darwin is not doing anything. It just exists
 // to make compilation possible.
