+++ pkg/datapath/sysctl_linux.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package datapath
 
 import (
 	"github.com/cilium/cilium/pkg/option"
